requirements.txt
gspread 
oauth2client 
python-telegram-bot
python-dotenv